+++ stream-chat-android/src/main/java/com/getstream/sdk/chat/utils/LlcMigrationUtils.java
@@ -3,6 +3,8 @@ package com.getstream.sdk.chat.utils;
 
 import android.text.TextUtils;
 
+import androidx.annotation.Nullable;
+
 import com.getstream.sdk.chat.R;
 import com.getstream.sdk.chat.model.AttachmentMetaData;
 import com.getstream.sdk.chat.model.ModelType;
@@ -15,7 +17,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import androidx.annotation.Nullable;
 import io.getstream.chat.android.client.logger.ChatLogger;
 import io.getstream.chat.android.client.models.Attachment;
 import io.getstream.chat.android.client.models.Channel;
@@ -65,8 +66,9 @@ public class LlcMigrationUtils {
         } else if (lastMessage == null) {
             return true;
         } else {
-            Date date = lastMessage.getCreatedAt() != null ? lastMessage.getCreatedAt() : lastMessage.getCreatedLocallyAt();
-            return myReadDate.getTime() > (date != null ? date.getTime() : 0);
+            Date lastMessageDate = lastMessage.getCreatedAt() != null ? lastMessage.getCreatedAt() : lastMessage.getCreatedLocallyAt();
+            boolean shouldBeMarkedRead = myReadDate.getTime() >= (lastMessageDate != null ? lastMessageDate.getTime() : 0);
+            return shouldBeMarkedRead;
         }
     }
 
@@ -94,9 +96,9 @@ public class LlcMigrationUtils {
 
     public static Date getLastActive(List<Member> members) {
         Date lastActive = new Date();
-        for (Member member: members) {
+        for (Member member : members) {
             if (member.getUser().getId() != ChatDomain.instance().getCurrentUser().getId()) {
-                if (member.getUser().getLastActive()!= null) {
+                if (member.getUser().getLastActive() != null) {
                     lastActive = member.getUser().getLastActive();
                 }
 
@@ -212,7 +214,7 @@ public class LlcMigrationUtils {
         return true;
     }
 
-    public static List<User> getOtherUsers( List<Member> members) {
+    public static List<User> getOtherUsers(List<Member> members) {
 
         List<User> result = new ArrayList<>();
 
@@ -337,7 +339,7 @@ public class LlcMigrationUtils {
         return lastMessage;
     }
 
-    public static User getCurrentUser(){
+    public static User getCurrentUser() {
         return ChatDomain.instance().getCurrentUser();
     }
 }